Keynote at DSAA 2025 in Birmingham
Title: The Art of the Pivot: Detecting Shifts in your Data Streams
Abstract: Our world is a continuous stream of data, but the most critical insights often lie not in the flow, but in the sudden, meaningful shifts. This talk is an accessible guide to changepoint detection, the powerful data science discipline of identifying abrupt, significant changes in the properties of a time series. We will explore how these “changepoints” are more than just anomalies—they are signals that can indicate a market crash, a machine failure, a fundamental shift in customer behaviour, or a pivotal moment in a patients treatment.
Through a series of real-world examples, this keynote will demystify the core concepts behind changepoint detection. We’ll discuss why it’s a crucial tool in the data science toolkit, and touch on the key methodologies used to find these hidden signals. By the end of this talk, you will gain a new perspective on analyzing data, recognizing that knowing when a change occurred is often the first step to unlocking its true value.